Fix type conversion error in BatchService
- Cast int to long for JobParameters.addLong() calls
- Fix compilation error: incompatible types int to Long
- Project now compiles successfully

+++ src/main/java/com/example/tcgbackend/service/BatchService.java
@@ -21,8 +21,8 @@ public class BatchService {
         JobParameters jobParameters = new JobParametersBuilder()
                 .addLong("time", System.currentTimeMillis())
                 .addString("tcgType", tcgType.name())
-                .addLong("startIndex", startIndex)
-                .addLong("endIndex", endIndex)
+                .addLong("startIndex", (long) startIndex)
+                .addLong("endIndex", (long) endIndex)
                 .toJobParameters();
 
         jobLauncher.run(importCardsJob, jobParameters);
